Split implementation of ModuleFactory in two
